server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-socketio==5.11.1
openai==1.12.0